+++ src/compute_global_atom.cpp
@@ -483,7 +483,7 @@ void ComputeGlobalAtom::compute_peratom()
 
 double ComputeGlobalAtom::memory_usage()
 {
-  double bytes = nmax*nvalues * sizeof(double);
+  double bytes = (double)nmax*nvalues * sizeof(double);
   bytes += (double)nmax * sizeof(int);                    // indices
   if (varatom) bytes += (double)nmax * sizeof(double);    // varatom
   bytes += (double)maxvector * sizeof(double);            // vecglobal
